editor/requirements.txt
-i https://pypi.org/simple

anyio==3.6.2

click==8.1.3

colorama==0.4.6

dnspython==2.3.0

fastapi==0.95.0

h11==0.14.0

httptools==0.5.0

idna==3.4

motor==3.1.1

pymongo==4.3.3

python-dotenv==1.0.0

pyyaml==6.0

sniffio==1.3.0

starlette==0.26.1

typing-extensions==4.5.0

uvicorn[standard]==0.21.1

watchfiles==0.18.1

websockets==10.4

